research/teanga/historical-document-analysis/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "historical-document-analysis"
version = "0.1.0"
description = "Multi-Modal Deep Learning for Historical Document Analysis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Isaac Godfried"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "transformers>=4.4.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "albumentations>=1.3.0",
    "evaluate>=0.4.0",
    "requests>=2.28.0",
    "tokenizers>=0.13.0",
    "evaluate",
    "sentencepiece>=0.1.99",
    "python-doctr",
    "google-generativeai",
    "notebook",
    "fastapi",
    "accelerate",
    "python-Levenshtein",
    "matplotlib",
    "openai",
    "google-cloud-storage",
    "jinja2",
    "networkx",
    "pandas",
    "plotly",
    "protobuf",
    "pyparsing",
    "pytz",
    "scikit-learn",
    "scipy",
    "selenium",
    "six",
    "sympy",
    "tornado",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12",
    "black>=21.5b2",
    "isort>=5.9.1",
    "flake8>=3.9.2",
    "mypy>=0.812",
]

[project.urls]
"Homepage" = "https://github.com/AIStream-Peelout/historical-document-analysis"
"Bug Tracker" = "https://github.com/AIStream-Peelout/historical-document-analysis/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"